# New engine implementation and performance optimizations