--- a/serve/mlc_serve/engine/__init__.py
+++ b/serve/mlc_serve/engine/__init__.py
@@ -4,6 +4,7 @@ from .base import (
     DebugOptions,
     FinishReason,
     InferenceEngine,
+    ScopedInferenceEngine,
     InferenceStepResult,
     Request,
     RequestId,