--- a/serve/mlc_serve/engine/base.py
+++ b/serve/mlc_serve/engine/base.py
@@ -121,6 +121,11 @@ class InferenceEngine:
         The output will contain empty delta and finish reason `cancelled`.
         """
 
+    def has_pending_requests(self) -> bool:
+        """
+        Check if there is pending requests in the engine.
+        """
+
     def wait_for_request(self, timeout_seconds=None) -> bool:
         """
         Block until there is request to process.
@@ -138,3 +143,28 @@ class InferenceEngine:
         If the engine has no requests in the queue, `step` will return immediately with
         an empty `InferenceStepResult.outputs`.
         """
+
+
+class ScopedInferenceEngine(InferenceEngine):
+    def start(self):
+        pass
+
+    def stop(self):
+        pass
+
+
+@dataclass
+class RequestState:
+    """
+    The internal state of request in the InferenceEngine.
+    """
+
+    request_id: RequestId
+    token_ids: list[int]
+    output_text: str
+    prompt_len: int
+    next_start_position: int
+    sampling_params: SamplingParams
+    stopping_criteria: StoppingCriteria
+    debug_options: DebugOptions
+    is_ended: bool = False