* New StagingInferenceEngine
  * It executes the token id generation loop in a separate process to make the loop tight. 
  * Tokenization and detokenization is done in the main process.
  * Basically this pipelines the prefill/decode and tokenization/detokenization. This is especially beneficial for large batch on small model (or big model with multiple GPUs).
* Separate Prefill and Decode
* Lower the default min_decode_steps
* Add a new param `max_decode_steps`

`StagingInferenceEngine` **brings 20% to 30% throughput improvement** on LLaMA 7B fp16 with single A100 80GB, compared to the old `SynchronousInferenceEngine`.

Using `python serve/benchmarks/benchmark_throughput.py --local-id llama-2-7b-chat-hf-q0f16 --dataset ~/ShareGPT_V3_unfiltered_cleaned_split.json  --max-num-batched-tokens 2560 --max-input-len 512 --num-prompts=1000`,
```
SynchronousInferenceEngine:
Throughput: 8.84 requests/s, 4225.16 tokens/s

StagingInferenceEngine:
Throughput: 10.62 requests/s, 5080.96 tokens/s
```

Under the context of serving API, server is started with 
```
python -m mlc_serve --local-id llama-2-7b-chat-hf-q0f16 --port=8081 --max-num-batched-tokens 2560 --max-input-len=256 --use-staging-engine
```
benchmarked with 200 concurrent requests
```
python3 -m tidallm.run --executor "shared-iterations" --vus 200 --iterations 800 --randomize-max-prompt-length 200 --randomize-max-output-len 200 --model llama-2-7b-chat --endpoint "http://0.0.0.0:8081/v1/chat/completions"
```
Result: 
```
SynchronousInferenceEngine:
Time to first token: 5.05, Toks/sec: 16.82
Latency per token: 59.4511540111433 ms
Throughput: 5866.486260186936 tokens / sec

StagingInferenceEngine:
Time to first token: 3.70, Toks/sec: 22.34
Latency per token: 44.754588647499375 ms
Throughput: 7809.889487796473 tokens / sec
```

The performance improvement would be less significant on larger models.

Caveats:
1. Duplicate code between `StagingInferenceEngine` and `SynchronousEngine`. We can either extract them, if we decide to keep two engine implementations, or delete `SynchronousEngine`, if we decide that `StagingInferenceEngine` is superior in all aspects.
2. The error handling in the engine isn't quite good yet. I've seen `probability tensor contains torch fill nanaeither inf, nan or element < 0` quite often during benchmark under certain kinds of engine parameters. Currently the engine just throws the exception and everything stops. It would be better if it creates error response for all requests in the batch, and continue processing remaining requests.

cc @masahi